OSIS-1698 Move generator to business
